requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
flask-migrate==4.0.4
pytube==15.0.0
yt-dlp==2023.7.6
assemblyai==0.17.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
bcrypt==4.0.1
ratelimit==2.2.1
